EnsembleInference

EnsembleInference.DiffusionNormal

EnsembleInference.DiracDelta

EnsembleInference.Haar

EnsembleInference.inversion

EnsembleInference.kron_delta

EnsembleInference.representation_block

EnsembleInference.representation_block

EnsembleInference.DiffusionNormal — Type

DiffusionNormal{Tμ,TΣ,Te,TB,TD,TM<:Manifolds.AbstractGroupManifold} <:
Manifolds.MPointDistribution{TM}

Constructor

DiffusionNormal(
    M::Manifolds.AbstractGroupManifold,
    μ,
    Σ,
    e = Manifolds.Identity(M),
    basis = Manifolds.DefaultOrthogonalBasis(),
    direction = Manifolds.LeftAction()
)

Construct a DiffusionNormal distribution on a group manifold $M$, with points represented with the same type as the identity element $e$, which may be user-specified. The mode of the distribution is at $μ$, and $Σ$ is a diffusion matrix applied to the coordinates of the Lie algebra $𝔪 = Tₑ M$ using the orthogonal basis.

EnsembleInference.DiracDelta — Type

DiracDelta{T,TM<:Manifolds.Manifold} <: Manifolds.MPointDistribution{TM}

Dirac distribution around a point $p$, that is, a distribution whose support is a single point $p$ on a manifold $M$.

Constructor

DiracDelta(M::Manifold, p)

EnsembleInference.Haar — Type

Haar{T,TM<:Manifolds.AbstractGroupManifold} <: Manifolds.MPointDistribution{TM}

Haar (invariant, i.e. uniform) measure on group manifold $M$.

Constructor

Haar(M::AbstractGroupManifold, p)

Constructs a Haar distribution on the group manifold $M$ with points represented like the provided point $p$, used for random sampling when implemented for compact groups.

EnsembleInference.inversion — Function

inversion(d::Distribution)

Given a distribution $d$ of elements $g ∈ G$, where $G$ is a group, the inversion of $d$ is the corresponding distribution of $g^{-1}$.

EnsembleInference.kron_delta — Method

kron_delta(i, j) -> Bool

The Kronecker delta function

\[δ_{i,j} = \begin{cases} 0 & \text{if } i ≠ j\\ 1 & \text{if } i = j \end{cases}\]

EnsembleInference.representation_block — Method

representation_block(::typeof(so3), X, ℓ)

Compute the size (2ℓ+1, 2ℓ+1) block $ℓ$ of the representation of the element $X$ of the element of the Lie algebra $𝔰𝔬(3)$.

The representation is defined as

\[u^ℓ(X) = \lim_{t → 0} \frac{d}{dt} U^ℓ(\exp(t X)),\]

where $U^ℓ(R)$ is the representation of the rotation $R$, and $\exp(R)$ is the group exponential map on $\mathrm{SO}(3)$.

EnsembleInference.representation_block — Method

representation_block(::typeof(SO3), p, ℓ) -> Matrix{Complex}

Compute the size (2ℓ+1, 2ℓ+1) block $ℓ$ of the representation of the rotation $p$.

This block is identical to the $D^ℓ$ block of the Wigner $D$-matrix using the conventions of [Varshalovich1988].

Examples

To create the "little-$d$" block $d^ℓ(\frac{π}{4}) = D^ℓ(0, \frac{π}{4}, 0)$,

julia> using EnsembleInference, Rotations

julia> EnsembleInference.representation_block(SO3, RotZXZ(0, π/2, 0), 1)
3×3 Matrix{ComplexF64}:
  0.5+0.0im              0.0-0.707107im  -0.5+0.0im
  0.0-0.707107im  1.2175e-16+0.0im        0.0-0.707107im
 -0.5+0.0im              0.0-0.707107im   0.5+0.0im

julia> EnsembleInference.representation_block(SO3, RotZXZ(0, π/2, 0), 2)
5×5 Matrix{ComplexF64}:
      0.25+0.0im   0.0-0.5im          -0.612372+0.0im           0.0+0.5im               0.25+0.0im
       0.0-0.5im  -0.5-0.0im                0.0-3.24998e-16im  -0.5-0.0im                0.0+0.5im
 -0.612372-0.0im   0.0-2.41689e-16im       -0.5-0.0im           0.0-3.62534e-16im  -0.612372-0.0im
       0.0+0.5im  -0.5+0.0im                0.0-2.03875e-16im  -0.5+0.0im                0.0-0.5im
      0.25+0.0im   0.0+0.5im          -0.612372+0.0im           0.0-0.5im               0.25+0.0im

Varshalovich1988Varshalovich, D. A., Moskalev, A. N. & Khersonskii, V. K. Quantum Theory of Angular Momentum. 1988. World Scientific, Singapore. ISBN: 9971-5-0107-4.