EnsembleInference

EnsembleInference.DiffusionNormal

EnsembleInference.Dirac

EnsembleInference.Haar

EnsembleInference.inversion

EnsembleInference.DiffusionNormal — Type

DiffusionNormal{Tμ,TΣ,Te,TB,TD,TM<:Manifolds.AbstractGroupManifold} <:
Manifolds.MPointDistribution{TM}

Constructor

DiffusionNormal(
    M::Manifolds.AbstractGroupManifold,
    μ,
    Σ,
    e = Manifolds.identity(M, μ),
    basis = Manifolds.DefaultOrthogonalBasis(),
    direction = Manifolds.LeftAction()
)

Construct a DiffusionNormal distribution on a group manifold $M$, with points represented with the same type as the identity element $e$, which may be user-specified. The mode of the distribution is at $μ$, and $Σ$ is a diffusion matrix applied to the coordinates of the Lie algebra $𝔪 = Tₑ M$ using the orthogonal basis.

EnsembleInference.Dirac — Type

Dirac{T,TM<:Manifolds.Manifold} <: Manifolds.MPointDistribution{TM}

Dirac distribution around a point $p$, that is, a distribution whose support is a single point $p$ on a manifold $M$.

Constructor

Dirac(M::Manifold, p)

EnsembleInference.Haar — Type

Haar{T,TM<:Manifolds.AbstractGroupManifold} <: Manifolds.MPointDistribution{TM}

Haar (invariant, i.e. uniform) measure on group manifold $M$.

Constructor

Haar(M::AbstractGroupManifold, p)

Constructs a Haar distribution on the group manifold $M$ with points represented like the provided point $p$, used for random sampling when implemented for compact groups.

EnsembleInference.inversion — Function

inversion(d::Distribution)

Given a distribution $d$ of elements $g ∈ G$, where $G$ is a group, the inversion of $d$ is the corresponding distribution of $g^{-1}$.